domain: move Permission and ValidationResults to domain crate
