+++ crates/domain/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod access_control;
 pub mod requests;
 pub mod schema;
 pub mod types;
