+++ server/src/infra/graphql/query.rs
@@ -786,16 +786,14 @@ impl<Handler: BackendHandler> AttributeValue<Handler> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::infra::{
-        access_control::{Permission, ValidationResults},
-        test_utils::{setup_default_schema, MockTestBackendHandler},
-    };
+    use crate::infra::test_utils::{setup_default_schema, MockTestBackendHandler};
     use chrono::TimeZone;
     use juniper::{
         execute, graphql_value, DefaultScalarValue, EmptyMutation, EmptySubscription, GraphQLType,
         RootNode, Variables,
     };
     use lldap_domain::{
+        access_control::{Permission, ValidationResults},
         schema::{AttributeList, Schema},
         types::{AttributeName, AttributeType, LdapObjectClass},
     };
